fix(web): replace footer
