+++ packages/web/components/Footer/styles.js
@@ -1,114 +1,182 @@
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 
 export const StyledFooter = styled.footer`
 	width: 100%;
 `;
 
 export const FooterHeader = styled.div`
-	padding: 0 3rem;
-	width: 100%;
-	background-color: ${({ theme }) => theme.colors.secondary};
+	${({ theme: { colors, screens } }) => css`
+		padding: 0 3rem;
+		background-color: ${colors.primary};
+		display: flex;
 
-	@media (max-width: ${({ theme }) => theme.screens.medium}px) {
-		padding: 0;
-	}
+		@media (max-width: ${screens.large}px) {
+			padding: 0;
+		}
+	`}
 `;
 
 export const FooterHeaderContainer = styled.div`
-	max-width: ${({ theme }) => theme.metrics.containerWidth}rem;
-	width: 100%;
-	height: 10rem;
-	margin: 0 auto;
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
+	${({ theme: { metrics, screens } }) => css`
+		max-width: ${metrics.containerWidth}rem;
+		width: 100%;
+		height: 8rem;
+		margin: 0 auto;
+		display: flex;
+		align-items: center;
+		justify-content: space-between;
 
-	> a {
-		padding: 1rem 0;
-		img {
-			width: 100%;
+		width: ${`${screens.large}px`};
+		max-width: 100%;
+		margin: 0 auto;
+
+		> a {
+			padding: 1rem 0;
+			img {
+				width: 100%;
+			}
 		}
+	`}
+`;
 
-		@media (max-width: ${({ theme }) => theme.screens.medium}px) {
+export const FooterText = styled.p`
+	${({ theme: { colors, screens } }) => css`
+		color: ${colors.white};
+		font-weight: 500;
+		font-size: 3.2rem;
+		line-height: 100%;
+
+		@media (max-width: ${screens.large}px) {
 			display: none;
 		}
-	}
+	`}
 `;
 
 export const FooterIconsList = styled.ul`
-	display: flex;
-	justify-content: space-between;
-	width: 40%;
-
-	@media (max-width: ${({ theme }) => theme.screens.medium}px) {
-		width: 100%;
+	${({ theme: { screens } }) => css`
+		display: flex;
 		justify-content: space-around;
-	}
+		height: 100%;
+
+		@media (max-width: ${screens.large}px) {
+			width: 100%;
+		}
+	`}
 `;
 
 export const FooterIconsListItem = styled.li`
-	font-size: 2.4rem;
+	${({ theme: { colors, sizes, screens } }) => css`
+		font-size: 2.4rem;
+		height: 100%;
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		width: 25%;
+		padding: 0 5rem;
+		border-left: 1px solid ${colors.white};
 
-	a {
-		color: ${({ theme }) => theme.colors.white};
-		:hover {
-			color: ${({ theme }) => theme.colors.primary};
+		&:last-of-type {
+			border-right: 1px solid ${colors.white};
 		}
-	}
 
-	svg {
-		height: ${({ theme }) => theme.sizes.defaultIcon}rem;
-		width: ${({ theme }) => theme.sizes.defaultIcon}rem;
-	}
+		a {
+			color: ${colors.white};
+		}
+
+		svg {
+			height: ${sizes.defaultIcon}rem;
+			width: ${sizes.defaultIcon}rem;
+		}
+
+		@media (max-width: ${screens.large}px) {
+			padding: 0;
+			border-left: 0;
+
+			& + li {
+				border-left: 1px solid ${colors.white};
+			}
+
+			&:last-of-type {
+				border-right: 0;
+			}
+		}
+	`}
 `;
 
 export const SiteInfo = styled.div`
-	padding: 9rem 3rem;
-	background-color: ${({ theme }) => theme.colors.border};
-
-	@media (max-width: ${({ theme }) => theme.screens.medium}px) {
-		text-align: center;
-		padding: 3rem;
-	}
+	${({ theme: { colors } }) => css`
+		padding: 3.2rem 2.2rem;
+		background-color: ${colors.secondary};
+	`}
 `;
 
 export const SiteInfoContainer = styled.div`
-	max-width: ${({ theme }) => theme.metrics.containerWidth}rem;
-	width: 100%;
-	margin: 0 auto;
+	${({ theme: { screens } }) => css`
+		width: ${`${screens.large}px`};
+		max-width: 100%;
+		margin: 0 auto;
 
-	display: grid;
-	grid-template-columns: repeat(auto-fit, minmax(22.5rem, 1fr));
-	grid-gap: 2rem 0rem;
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(22.5rem, 1fr));
+		row-gap: 1rem;
+		grid-gap: 2rem 0rem;
+		column-gap: 1rem;
+		row-gap: 1rem;
 
-	@media (max-width: ${({ theme }) => theme.screens.medium}px) {
-		grid-gap: 3.5rem 0rem;
-	}
+		@media (max-width: ${screens.huge}px) {
+			> div:first-child {
+				grid-column: 1 / -1;
+			}
+		}
+
+		@media (max-width: ${screens.large}px) {
+			grid-template-columns: repeat(auto-fit, minmax(13.7rem, 1fr));
+			grid-gap: 2.4rem 0;
+		}
+
+		img {
+			max-width: 100%;
+		}
+	`}
 `;
 
 export const SiteInfoListTitle = styled.h4`
-	font-weight: 500;
-	text-transform: uppercase;
-	margin-bottom: 3rem;
+	${({ theme: { colors, screens } }) => css`
+		font-weight: 500;
+		margin-bottom: 0.8rem;
+		font-size: 2.4rem;
+		color: ${colors.white};
 
-	@media (max-width: ${({ theme }) => theme.screens.medium}px) {
-		margin-bottom: 1rem;
-	}
+		@media (max-width: ${screens.large}px) {
+			font-size: 1.6rem;
+		}
+	`}
 `;
 
 export const SiteInfoListItem = styled.li`
-	padding: 1rem 0;
-	width: 100%;
-	color: ${({ theme }) => theme.colors.secondary};
-	letter-spacing: 0.1rem;
+	${({ theme: { colors, screens } }) => css`
+		width: 100%;
+		color: ${colors.secondary};
+		padding-bottom: 0.8rem;
 
-	a {
-		font-weight: 400;
-		color: ${({ theme }) => theme.colors.secondary};
-		:hover {
-			color: ${({ theme }) => theme.colors.darkGreen};
+		& + li {
+			padding-top: 0.8rem;
 		}
-	}
+
+		a {
+			font-weight: 500;
+			font-size: 1.6rem;
+			color: ${colors.white};
+
+			:hover {
+				color: ${colors.primary};
+			}
+
+			@media (max-width: ${screens.large}px) {
+				font-size: 1.4rem;
+			}
+		}
+	`}
 `;
 
 export const SiteSocket = styled.div`
@@ -133,6 +201,10 @@ export const SiteSocketContainer = styled.div`
 	width: 100%;
 	margin: 0 auto;
 
+	> div {
+		width: 50%;
+	}
+
 	span {
 		font-weight: bold;
 	}
@@ -143,6 +215,7 @@ export const SiteSocketContainer = styled.div`
 		div {
 			text-align: center;
 			margin-bottom: 3rem;
+			width: 100%;
 		}
 	}
 `;
