+++ packages/web/components/_Layouts/LandingPage.js
@@ -2,14 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Head from '../head';
 import { Header } from '../LandingPage/Header';
-import { Footer } from '../LandingPage/Footer';
+import { Footer } from '../Footer';
 
 const LayoutLandingPage = ({ children }) => (
 	<>
 		<Head title="Plataforma Sabiá" description="A Plataforma do Semi-Árido Brasileiro" />
 		<Header />
 		{children}
-		<Footer />
+		<Footer isAbout />
 	</>
 );
